+++ harvest/experts-client/lib/query/expert/bind.rq
@@ -1,11 +1,11 @@
 prefix : <http://oapolicy.universityofcalifornia.edu/vocab#>
-prefix person: <http://experts.ucdavis.edu/person/>
+prefix expert: <http://experts.ucdavis.edu/expert/>
 PREFIX harvest_oap: <http://oapolicy.universityofcalifornia.edu/>
 select distinct (?username as ?USERNAME__) ?filename where {
   graph harvest_oap: {
     [] :username ?username.
     #    bind(concat("person/",md5(?username)) as ?id)
     bind(md5(?username) as ?id)
-    bind(concat(?id,".jsonld.json") as ?filename)
+    bind(concat("expert/",?id,".jsonld.json") as ?filename)
   }
 }
