+++ services/base-service/models/authorship/index.js
@@ -1,7 +1,7 @@
 const RelationshipModel = require('./model.js');
 module.exports = {
   api : require('./api.js'),
-  model : new RelationshipModel(),
+  model : new AuthorshipModel(),
   schema : require('../experts/schema/minimal.json'),
   transform: require('../experts/transform.js')
 }
