+++ services/base-service/models/authorship/model.js
@@ -1,15 +1,14 @@
 // Can use this to get the fin configuration
 //const {config} = require('@ucd-lib/fin-service-utils');
-const ExpertsModel = require('../experts/model.js');
+const BaseModel = require('../base/model.js');
 
 /**
- * @class RelationshipModel
+ * @class AuthorshipModel
  * @description Base class for Aggie Experts data models.
  */
-class RelationshipModel extends ExpertsModel {
+class AuthorshipModel extends BaseModel {
 
   static types = [
-    "http://schema.library.ucdavis.edu/schema#Relationship",
     "http://schema.library.ucdavis.edu/schema#Authorship",
     "http://vivoweb.org/ontology/core#Authorship",
     "http://vivoweb.org/ontology/core#Relationship"
@@ -31,83 +30,71 @@ class RelationshipModel extends ExpertsModel {
       '@type': node['@type'],
       '@graph': [node]
     };
+    return doc;
+  }
 
-    /**
+  /**
    * @method update
    * @description Update Elasticsearch with the given data.
    */
 
-  async update(jsonld) {
-    console.log(`RelationshipModel.update(${jsonld['@id']})`);
-
-    const root_node= this.get_expected_model_node(jsonld);
+  async update(transformed) {
+    const root_node= this.get_expected_model_node(transformed);
     const doc = this.promote_node_to_doc(root_node);
-    doc['@graph']=jsonld['@graph'];
+    logger.info(`AuthorshipModel.update(${doc['@id']})`);
     await this.update_or_create_main_node_doc(doc);
 
-// Start Here
-    let have={};
+    const have_part={};
     // Get the work and the Person via the Authorship.relates
     const personModel= await this.get_model('person');
     const workModel= await this.get_model('work');
+    // The root_node is the Authorship node, pointers to Person and Work (which is now another node)
     for(let i=0; i<root_node?.relates?.length || 0; i++) {
       let relates = root_node.relates[i];
       try {
         // Is this a Person?
-        // console.log(`get ${relates}`);
         let related = await personModel.client_get(relates);
-        // console.log(`got ${related}`);
-        related=this.get_main_graph_node(related);
-        // This part is removed when we split the indices
         let type = this.experts_node_type(related);
         if (type !== 'Person') {
-          throw new Error(`RelationshipModel.update(${jsonld['@id']}) - ${relates} is not a Person`);
+          throw new Error(`AuthorshipModel.update(${doc['@id']}) - ${relates} is not a Person`);
         }
-        have['Person'] = {id:relates,node:related };
+        have_part['Person'] = {id:relates,node:related };
       } catch (e) {
         try {
-          // Is this a Work?
-          // get node of work type
-          let related = await workModel.client_get(relates);
-          related=this.get_main_graph_node(related);
+          // Is this included Work?
+          let related = workModel.get_expected_model_node(transformed);
           let type = this.experts_node_type(related);
+          if (relates !== related[0])
+            throw new Error(`AuthorshipModel.update(${relates} not included in doc`);
           if (type !== 'Work') {
-            throw new Error(`RelationshipModel.update(${jsonld['@id']}) - ${relates} is not a Work`);
+            throw new Error(`AuthorshipModel.update(${relates} is not a Work`);
           }
-          have['Work'] = {id:relates,node:related };
+          have_part['Work'] = {id:relates,node:related };
         } catch(e) {
-          console.log(`RelationshipModel.update(${jsonld['@id']}) - ${relates} is not a Person or Work`);
+          logger.warn(`AuthorshipModel.update(${relates} is not a Person or Work`);
         }
       }
-      if (have.Person && have.Work) {
+      if (have_part.Person && have_part.Work) {
         // Add Work as snippet to Person
         // console.log(root_node);
         if (root_node['is-visible'] === true || root_node['is-visible'] === 'true') {
-          console.log(`${have.Person.id} <=> ${have.Work.id}`);
+          logger.info(`${have_part.Person.id} ==> ${have_part.Work.id}`);
           {
             const node = {
-              ...workModel.snippet(have.Work.node),
+              ...workModel.snippet(have_part.Work.node),
               ...this.snippet(root_node),
-              '@type': 'Authored'
+              '@type': 'Authored',
             };
             delete node.relates;
-            // console.log(`${have.Person.id} Authored ${have.Work.id}`);
-            await personModel.update_graph_node(have.Person.id,node,root_node['is-visible']);
-          }
-          {
-            const node = {
-              ...personModel.snippet(have.Person.node),
-              ...this.snippet(root_node),
-              '@type': 'Author'
-            };
-            delete node.relates;
+            // console.log(`${have_part.Person.id} Authored ${have_part.Work.id}`);
+            await personModel.update_graph_node(have_part.Person.id,node,root_node['is-visible']);
           }
         } else {
-          console.log(`${have.Person.id} X=X ${have.Work.id}`);
-          await personModel.delete_graph_node(have.Person.id,root_node);
+          logger.info(`${have_part.Person.id} !=> ${have_part.Work.id}`);
+          await personModel.delete_graph_node(have_part.Person.id,root_node);
         }
       }
     }
   }
 }
-module.exports = RelationshipModel;
+module.exports = AuthorshipModel;
