+++ services/base-service/models/base/schema/minimal.json
