+++ services/base-service/models/base/template/default.json
