+++ services/base-service/models/expert/api.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const {dataModels,logger} = require('@ucd-lib/fin-service-utils');
-const PersonModel = require('./model.js');
+const ExpertModel = require('./model.js');
 const utils = require('../utils.js')
 const {defaultEsApiGenerator} = dataModels;
 
@@ -9,7 +9,7 @@ router.get('/test', async (req, res) => {
   res.send('ok');
 });
 
-const model = new PersonModel();
+const model = new ExpertModel();
 module.exports = defaultEsApiGenerator(model, {router});
