+++ services/base-service/models/expert/index.js
@@ -1,7 +1,7 @@
-const PersonModel = require('./model.js');
+const ExpertModel = require('./model.js');
 module.exports = {
   api : require('./api.js'),
-  model : new PersonModel(),
+  model : new ExpertModel(),
   schema : require('../experts/schema/minimal.json'),
   transform: require('../experts/transform.js')
 }
