+++ services/base-service/models/expert/model.js
@@ -1,0 +1,125 @@
+// Can use this to get the fin configuration
+//const {config} = require('@ucd-lib/fin-service-utils');
+const BaseModel = require('../base/model.js');
+
+/**
+ * @class ExpertModel
+ * @description Base class for Aggie Experts data models.
+ */
+class ExpertModel extends BaseModel {
+
+  static types = [
+    "http://schema.library.ucdavis.edu/schema#Expert",
+  ];
+
+  constructor(name='expert') {
+    super(name);
+  }
+
+  /**
+   * @method snippet
+   * @description returns searchable snippet of a node
+   * by elasticsearch.
+   */
+  snippet(node) {
+    const snippet= ["identifier","orcidId","name","contactInfo"];
+
+    // Get only best contact info
+    if (node.contactInfo) {
+      let best=node.contactInfo.sort((a,b) => {
+        (a['rank'] || 100) - (b['rank'] || 100)})[0];
+      ['hasOrganizationalUnit','hasTitle','hasURL','rank'].forEach(x => delete best[x]);
+      node.contactInfo = [best];
+    }
+
+    // Now select some of the fields.
+    let s = {};
+    snippet.forEach((key) => {
+      if (node[key]) {
+        s[key] = node[key];
+      }
+    });
+    return s;
+  }
+
+  /**
+   * @method promote_node_to_doc
+   * @description Promotes some node fields to document fields
+   * @param {Object} node
+   * @returns {Object} : document
+   **/
+  promote_node_to_doc(node) {
+    const doc = {
+      "@id": node['@id'],
+      "@graph": [node]
+    };
+
+    doc["@type"] = "Expert";
+
+    // Order the vcards, and get the first one
+    let contact
+    let hasEmail=[];
+    let hasURL=[];
+    if (node["contactInfo"]) {
+      if (! Array.isArray(node["contactInfo"])) {
+        node["contactInfo"] = [ node["contactInfo"] ];
+      } else {
+        node["contactInfo"].sort((a,b)=>a["rank"]-b["rank"])
+      }
+      contact = node["contactInfo"]?.[0];
+      // get the hasURL
+      node["contactInfo"].forEach((info)=>{
+        if (info.hasEmail) {
+          hasEmail=hasEmail.concat(info.hasEmail);
+        }
+
+        if (info?.hasURL) {
+          hasURL=hasURL.concat(info.hasURL);
+        }
+      });
+    }
+
+    doc["contactInfo"] = {};
+
+    if (hasURL.length > 0) {
+      doc.contactInfo["hasURL"] = hasURL;
+    }
+
+    doc.contactInfo["hasEmail"] = hasEmail?.[0];
+
+    ["name","hasName","hasTitle","hasOrganizationalUnit"].forEach((key)=>{
+      if (contact[key]) {
+        doc.contactInfo[key] = contact[key];
+      }
+    });
+    if (doc.contactInfo.name) {
+      doc.name = framed.contactInfo.name;
+    }
+
+    return doc;
+  }
+
+  /**
+   * @method update
+   * @description Update Elasticsearch with the given data.
+   */
+
+  async update(transformed) {
+    const root_node= this.get_expected_model_node(transformed);
+    const doc = this.promote_node_to_doc(root_node);
+    console.log(`${this.constructor.name}.update(${doc['@id']})`);
+    await this.update_or_create_main_node_doc(doc);
+
+    const authorshipModel=await this.get_model('authorship');
+    const workModel=await this.get_model('work');
+
+    // Update all Works with this Person as well
+    let authorships= await authorshipModel.esMatchNode({ 'relates': doc['@id'] });
+
+    for (let i=0; i<authorships?.hits?.hits?.length || 0; i++) {
+      authorshipModel.update(authorships.hits.hits[i]._source);
+      workModel.update(authorships.hits.hits[i]._source);
+    }
+  }
+}
+module.exports = ExpertModel;
