--- services/base-service/models/experts/model.js
@@ -1,280 +1,0 @@
-// Can use this to get the fin configuration
-const {config, models } = require('@ucd-lib/fin-service-utils');
-const schema = require('./schema/minimal.json');
-const FinEsNestedModel = require('../fin-es-nested-model');
-
-/**
- * @class ExpertsModel
- * @description Base class for Aggie Experts data models.
- */
-class ExpertsModel extends FinEsNestedModel {
-
-  // Experts Model never matches
-  static types = [
-//    "http://schema.library.ucdavis.edu/schema#Person",
-//    "http://schema.library.ucdavis.edu/schema#Work",
-//    "http://schema.library.ucdavis.edu/schema#Authorship",
-//    "http://vivoweb.org/ontology/core#Authorship",
-//    "http://vivoweb.org/ontology/core#Grant",
-  ];
-
-  constructor(name='experts') {
-
-    super(name);
-    this.schema = schema;  // Common schema for all experts data models
-    this.transformService = "node";
-  }
-
-  /**
-   * @method is
-   * @description Determines if this model can handle the given file based on
-   * it's type.
-   */
-  is(id,types,workflows) {
-//    console.log('constructor', this.constructor, 'vs', types);
-    if (typeof types === 'string') types = [types];
-    types = types.filter(x => this.constructor.types.includes(x));
-    if (types.length === 0) {
-      console.log(`!${this.constructor.name}.is`);
-      return false;
-    }
-    console.log(`+${this.constructor.name}.is(${types.join(",")} is a valid type)`);
-    return true
-  }
-
-    /**
-   * @method get_nodes_by_type
-   * @description Return the node(s) in the graph that corresponds to the '@type'
-   * @param {String} doc :  document
-   * @param {Object or Array} type : type, or array of types to match
-   *
-   * @returns {Array} : nodes in the graph that match by type
-   **/
-  get_nodes_by_type(doc,want_types) {
-    const nodes = [];
-    if (typeof want_types === 'string') want_types = [want_types];
-    for(let i=0; i<doc['@graph'].length; i++) {
-      let node = doc['@graph'][i];
-      let types= node.["@type"];
-      types = types.filter(x => want_types.includes(x));
-      if (types.length > 0) {
-        nodes.push(node);
-      }
-    }
-    return nodes;
-  }
-
-  /**
-   * @method get_expected_node_by_type
-   * @description Return single/expected node in the graph that corresponds to the '@type'
-   * @param {String} doc :  document
-   * @param {Object or Array} type : type, or array of types to match
-   *
-   * @returns {Object} : node in the graph matched by type
-   * @error {Error} : if no node in the graph matches the type, or if more than one node matches the type
-   **/
-  get_expected_node_by_type(doc,want_types) {
-    const nodes = this.get_nodes_by_type(doc,want_types);
-    if (nodes.length === 0) {
-      throw new Error(`get_expected_node_by_type: Unable to find node in ${doc['@id']} with type ${want_types.join(",")}`);
-    }
-    if (nodes.length > 1) {
-      throw new Error(`get_expected_node_by_type: Found multiple nodes in ${doc['@id']} with type ${want_types.join(",")}`);
-    }
-    return nodes[0];
-  }
-
-  /**
-   * @method get_expected_model_node
-   * @description Return single/expected node in the graph that corresponds to the '@type' of the model
-   * @param {Object} doc :  document
-   * @returns {Object} : node in the graph matched by type
-   * @error {Error} : if no node in the graph matches the type, or if more than one node matches the type
-   **/
-  get_expected_model_node(doc) {
-    const types = this.constructor.types;
-    return this.get_expected_node_by_type(doc,types);
-  }
-
-  /**
-   * @method experts_node_type
-   * @description Get the experts node type for the given type
-   * @param {String} node
-   */
-  experts_node_type(node) {
-    const Types = ['Person','Work','Grant','Relationship','Authorship'];
-    let types;
-    // Look for valid type in index
-    types=node['@type'];
-    // console.log(`experts_node_type: ${types}`);
-    if (typeof types === 'string') types = [types];
-    types = types.filter(x => Types.includes(x));
-    if (types.length > 1)
-      throw new Error(`update: ${jsonld['@id']} has multiple types: ${types.join(",")}`);
-    return types?.[0];
-  }
-
-  /**
-   * @method snippet
-   * @description returns searchable snippet of a node
-   * by elasticsearch.  Default is to return whole node.
-   */
-  snippet(node) {
-    delete node['_'];
-    return node;
-  }
-
-  /**
-   * @method promote_node_to_doc
-   * @description Promotes some node fields to document fields
-   * @param {Object} node
-   * @returns {Object} : document
-   **/
-  promote_node_to_doc(node) {
-    const doc = {
-      "@id": node['@id'],
-      name: node['name'],
-      "@graph": [node]
-    };
-    return doc;
-  }
-
-  /**
-   * @method update_or_create_doc_from_graph_node
-   * @description Update main @node of a document. Create if document does not exist.
-   * @param {String} document_id
-   * @param {Object} node_to_update
-   *
-   * @returns {Promise} : Elasticsearch response Promise
-   */
-  async update_or_create_doc_from_graph_node(node) {
-    const doc = this.promote_node_to_doc(node);
-    const roles = await this.getAccessRoles(doc);
-
-    return this.client.index({
-      index : this.writeIndexAlias,
-      id : doc['@id'],
-      body: {
-        ...doc,
-        roles: roles}
-    });
-  }
-
-
-  /** vvvv TEMPLATE SEARCH vvvv **/
-
-
-  /**
-   * @method common_parms
-   * @description fixup parms for template searches
-   * @returns object
-   */
-  common_parms(in_params) {
-    const params = {
-      size:10,
-      from:0,
-      ...in_params
-    }
-    // convert page to from if from is not set
-    if (params.from === 0 && params.page > 1) {
-      params.from = (params.page - 1) * params.size;
-    }
-    return params;
-  }
-
-  async verify_template(options) {
-    // Check if template exists, install if not
-    try {
-      const result = await this.client.getScript({id:options.id});
-      console.log(`render: template ${options.id} exists`);
-    } catch (err) {
-      console.log(`render: template ${options.id} does not exist`);
-      const template = require(`./template/${options.id}.json`);
-      const result = await this.client.putScript(template);
-    }
-    return true;
-  }
-
-  compact_search_results(results,params) {
-    const compact = {
-      params,
-      total: results.hits.total.value
-    }
-    const hits=[];
-    for (const hit of results.hits.hits) {
-      const source = hit._source;
-      const in_hits = hit?.inner_hits?.["@graph"].hits.hits;
-      const inner_hits = [];
-      if (in_hits) {
-        for (const in_hit of in_hits) {
-          inner_hits.push(in_hit._source);
-        }
-      }
-      source._inner_hits = inner_hits;
-      hits.push(source);
-    }
-    compact.hits = hits;
-    return compact;
-  }
-  /**
-   * @method render
-   * @description return an ES ready nested search using a template
-   * @returns string
-   */
-  async render(opts) {
-    const params = this.common_parms(opts.params);
-
-    const options = {
-      id: (opts.id)?opts.id:"default",
-      params
-    }
-    // Check if template exists, install if not
-    await this.verify_template(options);
-
-    const template = await this.client.renderSearchTemplate(options);
-    return template;
-  }
-
-  async search(opts) {
-
-    const params = this.common_parms(opts.params);
-
-    const options = {
-      id: (opts.id)?opts.id:"default",
-      index: "person-read",
-      params
-    }
-    // Check if template exists, install if not
-    await this.verify_template(options);
-
-    const res=await this.client.searchTemplate(options);
-    return this.compact_search_results(res,params);
-  }
-  /** ^^^^TEMPLATE SEARCH^^^^ **/
-
-
-
-  /**
-   * @method get_model
-   * @description returns a model for a given type
-   */
-  async get_model(model) {
-    const method=model+'Model';
-    if (!(method in this)) {
-      this.method = (await models.get(model)).model;
-//      console.log(`get_model: ${model} not found, loading...`);
-    }
-//    console.log(`get_model: ${model} found`,this.method);
-    return this.method;
-  }
-
-  async update(jsonld) {
-    throw new Error(`${this.constructor.name}.update(${jsonld['@id']}) not implemented`);
-  }
-
-  async remove(jsonld) {
-    throw new Error(`${this.constructor.name}.delete(${jsonld['@id']}) not implemented`);
-  }
-
-}
-module.exports = ExpertsModel;
