+++ services/base-service/models/index.js
@@ -1,6 +1,6 @@
 module.exports = {
   search: require('./search/index.js'),
-  person: require('./person/index.js'),
+  expert: require('./expert/index.js'),
   work: require('./work/index.js'),
   relationship: require('./relationship/index.js')
 }
